requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
lxml==4.0.0
olefile==0.44
Pillow==4.2.1
PyYAML==3.12
requests==2.18.4
urllib3==1.22
